get rid of ApiKeyManager
https://coveord.atlassian.net/browse/CDX-223
