--- packages/cli/src/lib/platform/apiKeyManager.ts
@@ -1,16 +1,0 @@
-import {AuthenticatedClient} from './authenticatedClient';
-
-export class ApiKeyManager {
-  public async createImpersonateApiKey(name: string) {
-    const authenticatedClient = new AuthenticatedClient();
-    const platformClient = await authenticatedClient.getClient();
-    return await platformClient.apiKey.create({
-      displayName: name,
-      description: 'Generated by the Coveo CLI',
-      enabled: true,
-      privileges: [
-        {targetDomain: 'IMPERSONATE', targetId: '*', owner: 'SEARCH_API'},
-      ],
-    });
-  }
-}
