+++ packages/cli/src/lib/platform/authenticatedClient.spec.ts
@@ -28,6 +28,7 @@ describe('AuthenticatedClient', () => {
   );
 
   const mockInitialize = jest.fn();
+  const mockCreate = jest.fn();
 
   mockConfig.mockImplementation(
     () =>
@@ -39,6 +40,9 @@ describe('AuthenticatedClient', () => {
     () =>
       (({
         initialize: mockInitialize,
+        apiKey: {
+          create: mockCreate as unknown,
+        },
       } as unknown) as PlatformClient)
   );
 
@@ -100,6 +104,21 @@ describe('AuthenticatedClient', () => {
     mockInitialize.mockReturnValueOnce(Promise.resolve());
     expect(await getAuthenticationStatus()).toBe(
       AuthenticationStatus.LOGGED_IN
+    );
+  });
+
+  it('should create an API key with impersonate privileges', async () => {
+    await new AuthenticatedClient().createImpersonateApiKey('my-key');
+
+    expect(mockCreate).toHaveBeenCalledWith(
+      expect.objectContaining({
+        displayName: 'my-key',
+        description: 'Generated by the Coveo CLI',
+        enabled: true,
+        privileges: [
+          {targetDomain: 'IMPERSONATE', targetId: '*', owner: 'SEARCH_API'},
+        ],
+      })
     );
   });
 });
