+++ packages/cli/src/lib/platform/authenticatedClient.ts
@@ -58,6 +58,18 @@ export class AuthenticatedClient {
     const orgs = await this.getAllOrgsUserHasAccessTo();
     return orgs.some((o) => o.id === org);
   }
+
+  async createImpersonateApiKey(name: string) {
+    const platformClient = await this.getClient();
+    return await platformClient.apiKey.create({
+      displayName: name,
+      description: 'Generated by the Coveo CLI',
+      enabled: true,
+      privileges: [
+        {targetDomain: 'IMPERSONATE', targetId: '*', owner: 'SEARCH_API'},
+      ],
+    });
+  }
 }
 
 export enum AuthenticationStatus {
